prep separation for intel rapl reader
